Create `map` Submodule
